requirements.txt
aiohttp==3.8.5
aiolimiter==1.0.0
aiosignal==1.3.1
openai==0.28.1
python-telegram-bot[webhooks]==20.5
pydantic==1.10.12
async-timeout==4.0.2
attrs==23.1.0
httpx==0.24.1
multidict==6.0.4
requests==2.31.0
python-dotenv==1.0.0
tiktoken==0.5.1
certifi==2023.7.22
idna==3.4
urllib3==2.0.4
yarl==1.9.2
